Add basic IPFIX packet parsing
Currently the code cowardly refuses to deal with variable length encoded
fields or any of the complex data types, skipping any template containing
them. Tested with `softflowd`.

Currently no field is "fixed up" such as timestamps; IPFIX unlike Netflow
does not include the uptime of the exporting device in each packet. However
testing with `softflowd` shows an initial option record is sent with the
systemInitTimeMilliseconds field which is what is needed for subsequent
calculations so maybe that should be cached per-device while packets are
received with some frequency. A device however may also include this field
as part of its normal templates.