Various fixes for IPFIX handling
1. Fixup absolute timestamp fields
Testing with OpenBSD's `pflow(4)` device a) worked and b) highlighted that it
exports using the flow{Start,End}Milliseconds fields which contain an absolute
timestamp. Fix this up to be a regular ISO 8601 timestamp and also add support
for the equivalent seconds, microseconds, and nanoseconds fields.

2. Update gem summary

3. Remove BinData tracing
Possibly causes stack errors having it enabled?

4. Improve invalid input handling
Based on similar changes to Netflow v9 code.