Update netflow.rb
Added 3 Netflow/IPFIX exporters are known to work. I tested them all with current version of plugin. 

Propose to use "?" to specify case if it's should work, not tested yet and "n" if it's not supported. Empty values needs to be removed as it's meaning is not clear (unknown or not supported).

For Juniper MX series we had bug with older software version - samplerd high CPU load and no flows generated. Details there http://www.gossamer-threads.com/lists/nsp/juniper/52052 , solution was to upgrade to Junos 13.3R9.